Abstract enable, include and exclude settings to a parent class
