+++ src/Core/Settings/AdditionalLanceSettings.cs
@@ -5,16 +5,7 @@ using Newtonsoft.Json;
 using System.Linq;
 
 namespace MissionControl.Config {
-  public class AdditionalLanceSettings {
-    [JsonProperty("Enable")]
-    public bool Enable { get; set; } = true;
-
-    [JsonProperty("IncludeContractTypes")]
-    public List<string> IncludeContractTypes { get; set; } = new List<string>();
-
-    [JsonProperty("ExcludeContractTypes")]
-    public List<string> ExcludeContractTypes { get; set; } = new List<string>();
-
+  public class AdditionalLanceSettings : AdvancedSettings {
     [JsonProperty("UseElites")]
     public bool UseElites { get; set; } = true;
 
@@ -35,33 +26,5 @@ namespace MissionControl.Config {
 
     [JsonProperty("DisableWhenMaxTonnage")]
     public DisableWhenMaxTonnage DisableWhenMaxTonnage { get; set; } = new DisableWhenMaxTonnage();
-
-    public List<string> GetValidContractTypes() {
-      List<string> validContracts = new List<string>();
-      bool useInclude = false;
-      bool useExclude = false;
-
-      if (ExcludeContractTypes.Count > 0) {
-        useExclude = true;
-      }
-
-      if (IncludeContractTypes.Count > 0) {
-        useExclude = false;
-        useInclude = true;
-      }
-
-      if (useInclude) {
-        validContracts.AddRange(IncludeContractTypes);
-      } else {
-        validContracts.AddRange(MissionControl.Instance.GetAllContractTypes());
-      }
-
-      if (useExclude) {
-        validContracts = validContracts.Except(ExcludeContractTypes).ToList();
-      }
-
-      Main.Logger.Log($"[ExtendedBoundaries.GetValidContractTypes] Valid contracts are '{string.Join(", ", validContracts.ToArray())}'");
-      return validContracts;
-    }
   }
 }
