+++ src/Core/Settings/ExtendedLancesSettings.cs
@@ -6,16 +6,7 @@ using BattleTech.Framework;
 using Newtonsoft.Json;
 
 namespace MissionControl.Config {
-  public class ExtendedLancesSettings {
-    [JsonProperty("Enable")]
-    public bool Enable { get; set; } = true;
-
-    [JsonProperty("IncludeContractTypes")]
-    public List<string> IncludeContractTypes { get; set; } = new List<string>();
-
-    [JsonProperty("ExcludeContractTypes")]
-    public List<string> ExcludeContractTypes { get; set; } = new List<string>();
-
+  public class ExtendedLancesSettings : AdvancedSettings {
     [JsonProperty("Autofill")]
     public bool Autofill { get; set; } = true;
 
@@ -48,34 +39,6 @@ namespace MissionControl.Config {
       }
 
       return lanceOverride.lanceDifficultyAdjustment;
-    }
-
-    public List<string> GetValidContractTypes() {
-      List<string> validContracts = new List<string>();
-      bool useInclude = false;
-      bool useExclude = false;
-
-      if (ExcludeContractTypes.Count > 0) {
-        useExclude = true;
-      }
-
-      if (IncludeContractTypes.Count > 0) {
-        useExclude = false;
-        useInclude = true;
-      }
-
-      if (useInclude) {
-        validContracts.AddRange(IncludeContractTypes);
-      } else {
-        validContracts.AddRange(MissionControl.Instance.GetAllContractTypes());
-      }
-
-      if (useExclude) {
-        validContracts = validContracts.Except(ExcludeContractTypes).ToList();
-      }
-
-      Main.Logger.Log($"[ExtendedLances.GetValidContractTypes] Valid contracts are '{string.Join(", ", validContracts.ToArray())}'");
-      return validContracts;
     }
   }
 }
