+++ src/Core/Settings/RandomSpawnsSettings.cs
@@ -4,42 +4,5 @@ using System.Linq;
 using Newtonsoft.Json;
 
 namespace MissionControl.Config {
-  public class RandomSpawnsSettings {
-    [JsonProperty("Enable")]
-    public bool Enable { get; set; } = true;
-
-    [JsonProperty("IncludeContractTypes")]
-    public List<string> IncludeContractTypes { get; set; } = new List<string>();
-
-    [JsonProperty("ExcludeContractTypes")]
-    public List<string> ExcludeContractTypes { get; set; } = new List<string>();
-
-    public List<string> GetValidContractTypes() {
-      List<string> validContracts = new List<string>();
-      bool useInclude = false;
-      bool useExclude = false;
-
-      if (ExcludeContractTypes.Count > 0) {
-        useExclude = true;
-      }
-
-      if (IncludeContractTypes.Count > 0) {
-        useExclude = false;
-        useInclude = true;
-      }
-
-      if (useInclude) {
-        validContracts.AddRange(IncludeContractTypes);
-      } else {
-        validContracts.AddRange(MissionControl.Instance.GetAllContractTypes());
-      }
-
-      if (useExclude) {
-        validContracts = validContracts.Except(ExcludeContractTypes).ToList();
-      }
-
-      Main.Logger.Log($"[RandomSpawnsSettings.GetValidContractTypes] Valid contracts are '{string.Join(", ", validContracts.ToArray())}'");
-      return validContracts;
-    }
-  }
+  public class RandomSpawnsSettings : AdvancedSettings { }
 }
